requirements.txt

PyMuPDF==1.26.5
PyMuPDF==1.26.5
pdfplumber==0.11.7
fastapi==0.119.0
uvicorn==0.37.0
python-dotenv==1.1.1
peewee==3.18.2
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
unstructured==0.18.15
unstructured-client==0.42.3
chromadb==1.1.1
huggingface-hub==0.35.3
langchain-huggingface==0.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
langchain-google-genai==2.0.10
sentence-transformers==5.1.1
